V1.0.2 added UNO_modbus_simple_chaining example

+++ examples/UNO_modbus_retry/UNO_modbus_retry.ino
@@ -89,7 +89,7 @@ void loop() {
     }
     if (!err) {
       Serial.print("response Len "); Serial.print(nbModbusMaster.getResponseBufferLength()); Serial.print("  ");
-      for (int i = 0; i < nbModbusMaster.getResponseBufferLength(); i++) {
+      for (size_t i = 0; i < nbModbusMaster.getResponseBufferLength(); i++) {
         nbModbusMaster.printHex(nbModbusMaster.getResponseBuffer(i), Serial); Serial.print(" ");
       }
       Serial.println();
